mapfusion:修复size()-1的问题 URL: http://qingluan.hozonauto.com/geelib/repoMergeDetail?repoId=algorithm/mapping&mergeId=1022

+++ modules/map_fusion/include/map_fusion/map_select/lane_line_delay.h
@@ -35,7 +35,7 @@ class Delay {
    */
   T GetAverageValue();
   /**
-   * @description: 该函数将当前数据存入， 
+   * @description: 该函数将当前数据存入，
    * @param {T&} input
    * @return {*}
    */
@@ -146,7 +146,7 @@ T Delay<T>::GetDelay(int num) {
   if (num < 1) {
     return deque_data_.front();
   }
-  if (num > (deque_data_.size() - 1)) {
+  if (num > (static_cast<int>(deque_data_.size()) - 1)) {
     return deque_data_.back();
   }
   return deque_data_.at(num);
